refact: split ark utils in order to jest test work
